icon/requirements.txt
pillow
numpy
svgpathtools